api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]~=0.116",
    "httpx~=0.28",
    "huggingface-hub~=0.34",
    "logtail-python~=0.3",
    "pydantic-settings~=2.10",
    "python-dotenv~=1.1",
    "sentry-sdk[fastapi]~=2.35",
    "torch~=2.8",
    "transformers~=4.56",
]

[dependency-groups]
dev = [
    "mypy~=1.17",
    "openapi-core~=0.19",
    "pytest-cov~=7.0",
    "pytest~=8.4",
    # If you change this, don't forget to change the pre-commit hook as well
    "ruff==0.12.12",
    "types-requests~=2.32",
]

[tool.ruff]
lint.select = ["E", "F", "I", "TID"]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
python_classes = ["*Test"]
python_functions = ["should_*"]
python_files = ["*_test.py"]
addopts = ["--cov=.", "--cov-report=xml"]

[tool.coverage.report]
fail_under = 100
show_missing = true
precision = 2

[tool.coverage.xml]
output = "coverage.xml"
